+++ src/components/Lexical/helpers/index.ts
@@ -278,3 +278,26 @@ export function getDomAttributes(dom: HTMLElement | undefined): HtmlHTMLAttribut
 
   return attributes;
 }
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function shallowEqual(obj1: any, obj2: any): boolean {
+  if (obj1 === obj2) return true;
+  if (!obj1 || !obj2) return false;
+
+  const keys1 = Object.keys(obj1);
+  const keys2 = Object.keys(obj2);
+
+  if (keys1.length !== keys2.length) return false;
+
+  for (const key of keys1) {
+    if (key === 'style') {
+      // - EN: Special handling for style objects.
+      // - CN: 特殊处理 style 对象。
+      if (!shallowEqual(obj1[key], obj2[key])) return false;
+    } else if (obj1[key] !== obj2[key]) {
+      return false;
+    }
+  }
+
+  return true;
+}