+++ CHANGELOG.md
@@ -8,6 +8,17 @@
 
 ### Features
 
+#### Lexical components
+
+- ✨ rename `getPropValue` to `getProp`
+- ✨ rename `setProps` to `updateProps`
+
+## 1.4.2
+
+2025-8-10
+
+### Features
+
 - ✨ `SelectNode` component now supports `onXXX` event props like `onClick`, `onMouseEnter`, etc.
 
 ## 1.4.1